Task 6. Express Fix errors
