+++ src/task6/cart/cart.repository.ts
@@ -1,42 +1,7 @@
 import { v4 as uuidv4 } from 'uuid';
-import { ProductEntity, product as bookProduct } from '../product/product.repository';
-import { UserEntity } from '../user/user.repository';
 
-export interface CartItemEntity {
-  product: ProductEntity;
-  count: number;
-}
-
-export interface CartEntity {
-  id: string;
-  userId: string;
-  isDeleted: boolean;
-  items: CartItemEntity[];
-}
-
-export interface TotalCartEntity {
-  cart: CartEntity;
-  total: number;
-}
-
-export interface AddCartItemEntity {
-  productId: string;
-  count: number;
-}
-
-const cartItem: CartItemEntity = {
-  product: bookProduct,
-  count: 2,
-};
-
-export const cart: CartEntity = {
-  id: '1434fec6-cd85-420d-95c0-eee2301a971d',
-  userId: '0fe36d16-49bc-4aab-a227-f84df899a6cb',
-  isDeleted: false,
-  items: [cartItem],
-};
-
-const carts: CartEntity[] = [cart];
+import { UserEntity } from '../user/user.entity';
+import { CartEntity, CartItemEntity, carts } from './cart.entity';
 
 export const getUserCart = (user: UserEntity): CartEntity | undefined => {
   return carts.find((cart: CartEntity) => cart.userId === user.id);
