+++ src/task6/order/order.service.ts
@@ -1,7 +1,9 @@
-import { CartEntity } from '../cart/cart.repository';
-import { addOrder, OrderDetailsEntity, OrderEntity, createOrderEntity } from './order.repository';
+import { addOrder, createOrderEntity } from './order.repository';
 import { deleteCart } from '../cart/cart.service';
-import { UserEntity } from '../user/user.repository';
+
+import { UserEntity } from '../user/user.entity';
+import { OrderDetailsEntity, OrderEntity } from './order.entity';
+import { CartEntity } from '../cart/cart.entity';
 
 export const createOrder = (user: UserEntity, cart: CartEntity, details: OrderDetailsEntity): OrderEntity => {
   const order: OrderEntity = createOrderEntity(cart, details);
