+++ src/task6/utils/models.ts
@@ -1,9 +1,9 @@
 import { Request } from 'express';
 
-import { CartEntity } from '../cart/cart.repository';
-import { UserEntity, UserToken } from '../user/user.repository';
-import { OrderEntity } from '../order/order.repository';
-import { ProductEntity } from '../product/product.repository';
+import { UserEntity, UserToken } from '../user/user.entity';
+import { ProductEntity } from '../product/product.entity';
+import { OrderEntity } from '../order/order.entity';
+import { CartEntity } from '../cart/cart.entity';
 
 export interface AuthenticatedRequest extends Request {
   user?: UserEntity;
